Update vectorize to maintain original aspect/size.
Previously, vectorizing an image would lose image
dimensions and aspect ratio if the traced image
did not extend to the edges of the vectorized
bitmap.  This patch keeps the dimensions of the
vectorized image.

Also, a trace that fails to find any pixels
(incorrectly set threshold) is now disposed of
properly.